LayoutTests/imported/w3c/web-platform-tests/webaudio/the-audio-api/the-gainnode-interface/no-dezippering-expected.txt


PASS # AUDIT TASK RUNNER STARTED. 
PASS Executing "test0" 
PASS Executing "test2" 
PASS Executing "test3" 
PASS Audit report 
PASS > [test0] Dezippering of GainNode removed 
FAIL X output[0:127]: Expected 1 for all values but found 128 unexpected values: 
	Index	Actual
	[0]	0.5
	[1]	0.5
	[2]	0.5
	[3]	0.5
	...and 124 more errors. assert_true: expected true got false
FAIL X output[128:]: Expected 0.5 for all values but found 128 unexpected values: 
	Index	Actual
	[768]	0
	[769]	0
	[770]	0
	[771]	0
	...and 124 more errors. assert_true: expected true got false
FAIL < [test0] 2 out of 2 assertions were failed. assert_true: expected true got false
PASS > [test2] Compare value setter and setValueAtTime 
PASS   .value setter output is identical to the array [-0.03759972006082535,-0.05467035621404648,-0.07155975699424744,-0.0882117748260498,-0.10457140952348709,-0.12058408558368683,-0.13619694113731384,-0.1513580083847046,-0.16601718962192535,-0.18012580275535583,-0.19363702833652496,-0.20650623738765717,-0.21869038045406342,-0.23014938831329346,-0.24084503948688507,-0.25074201822280884...]. 
PASS   .value setter output matches setValueAtTime output is true. 
PASS < [test2] All assertions passed. (total 2 assertions) 
PASS > [test3] Automation effects 
PASS   With modulation: .value setter output is identical to the array [-0.03759991005063057,-0.07845700532197952,-0.13329683244228363,-0.2007361203432083,-0.2788574993610382,-0.3652743101119995,-0.45721668004989624,-0.5516262650489807,-0.6452678442001343,-0.7348412275314331,-0.8171004056930542,-0.8889712691307068,-0.9476582407951355,-0.9907503128051758,-1.016305923461914,-1.0229262113571167...]. 
PASS   With modulation: .value setter output matches setValueAtTime output is true. 
PASS < [test3] All assertions passed. (total 2 assertions) 
FAIL # AUDIT TASK RUNNER FINISHED: 1 out of 3 tasks were failed. assert_true: expected true got false
